Exclude reflection only code from netfx on runtime
